+++ Plugins/Reqnroll.Microsoft.Extensions.DependencyInjection.ReqnrollPlugin/ServiceCollectionFinder.cs
@@ -3,58 +3,102 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
+using Reqnroll.Bindings.Discovery;
+using Reqnroll.Infrastructure;
 
 namespace Reqnroll.Microsoft.Extensions.DependencyInjection
 {
     public class ServiceCollectionFinder : IServiceCollectionFinder
     {
-        private readonly ITestRunnerManager testRunnerManager;
-        private (IServiceCollection, ScopeLevelType) _cache;
+        private readonly ITestRunnerManager _testRunnerManager;
+        private readonly IRuntimeBindingRegistryBuilder _bindingRegistryBuilder;
+        private readonly ITestAssemblyProvider _testAssemblyProvider;
+        private (IServiceCollection ServiceCollection, ScenarioDependenciesAttribute Attribute) _cache;
+        private readonly object _cacheLock = new();
 
-        public ServiceCollectionFinder(ITestRunnerManager testRunnerManager)
+        public ServiceCollectionFinder(ITestRunnerManager testRunnerManager, IRuntimeBindingRegistryBuilder bindingRegistryBuilder, ITestAssemblyProvider testAssemblyProvider)
         {
-            this.testRunnerManager = testRunnerManager;
+            _testRunnerManager = testRunnerManager;
+            _bindingRegistryBuilder = bindingRegistryBuilder;
+            _testAssemblyProvider = testAssemblyProvider;
+        }
+
+        public ServiceProviderLifetimeType GetServiceProviderLifetime()
+        {
+            if (_cache == default)
+            {
+                PopulateCache();
+            }
+
+            return _cache.Attribute.ServiceProviderLifetime;
         }
 
         public (IServiceCollection, ScopeLevelType) GetServiceCollection()
         {
-            if (_cache != default)
+            if (_cache == default)
             {
-                return _cache;
+                PopulateCache();
             }
 
-            var assemblies = testRunnerManager.BindingAssemblies;
-            foreach (var assembly in assemblies)
-            {
-                foreach (var type in assembly.GetTypes())
-                {
-                    foreach (MethodInfo methodInfo in type.GetMethods(BindingFlags.Static | BindingFlags.NonPublic | BindingFlags.Public))
-                    {
-                        var scenarioDependenciesAttribute = (ScenarioDependenciesAttribute)Attribute.GetCustomAttribute(methodInfo, typeof(ScenarioDependenciesAttribute));
+            return (_cache.ServiceCollection, _cache.Attribute.ScopeLevel);
+        }
 
-                        if (scenarioDependenciesAttribute != null)
+        private void PopulateCache()
+        {
+            if (_cache != default)
+            {
+                return;
+            }
+
+            lock (_cacheLock)
+            {
+                if (_cache != default)
+                {
+                    return;
+                }
+
+                var assemblies = _testRunnerManager.BindingAssemblies ?? _bindingRegistryBuilder.GetBindingAssemblies(_testAssemblyProvider.TestAssembly);
+                foreach (var assembly in assemblies)
+                {
+                    foreach (var type in assembly.GetTypes())
+                    {
+                        foreach (var methodInfo in type.GetMethods(BindingFlags.Static | BindingFlags.NonPublic | BindingFlags.Public))
                         {
-                            var serviceCollection = GetServiceCollection(methodInfo);
-                            if (scenarioDependenciesAttribute.AutoRegisterBindings)
+                            var scenarioDependenciesAttribute = (ScenarioDependenciesAttribute)Attribute.GetCustomAttribute(methodInfo, typeof(ScenarioDependenciesAttribute));
+
+                            if (scenarioDependenciesAttribute != null)
                             {
-                                AddBindingAttributes(assemblies, serviceCollection);
+                                var serviceCollection = GetServiceCollection(methodInfo);
+                                if (scenarioDependenciesAttribute.AutoRegisterBindings)
+                                {
+                                    AddBindingAttributes(assemblies, serviceCollection);
+                                }
+
+                                // If the ServiceProviderLifetime is Scenario, set the ScopeLevel to Scenario to match.
+                                if (scenarioDependenciesAttribute.ServiceProviderLifetime == ServiceProviderLifetimeType.Scenario)
+                                {
+                                    scenarioDependenciesAttribute.ScopeLevel = ScopeLevelType.Scenario;
+                                }
+
+                                _cache = (serviceCollection, scenarioDependenciesAttribute);
+                                return;
                             }
-                            return _cache = (serviceCollection, scenarioDependenciesAttribute.ScopeLevel);
                         }
                     }
                 }
+                var assemblyNames = assemblies.Select(a => a.GetName().Name).ToList();
+                throw new MissingScenarioDependenciesException(assemblyNames);
             }
-            var assemblyNames = assemblies.Select(a => a.GetName().Name).ToList();
-            throw new MissingScenarioDependenciesException(assemblyNames);
         }
 
         private static IServiceCollection GetServiceCollection(MethodInfo methodInfo)
         {
-            var serviceCollection = methodInfo.Invoke(null, null);
-            if(methodInfo.ReturnType == typeof(void))
+            if (methodInfo.ReturnType == typeof(void))
             {
                 throw new InvalidScenarioDependenciesException("the method doesn't return a value.");
             }
+
+            var serviceCollection = methodInfo.Invoke(null, null);
 
             if (serviceCollection == null)
             {